logic moved from emotion-records-page to searchBar; searchBar is nor rendered on emotions-records-page instead in EmotionRecordsList component;  messages adjusted when emotionEntries empty and searchResults empty, messages are now shown in emotion-records page (when emotionEntries are empty) and in EmotionRecordsList (when search results are empty);
